// src/types/configuration.ts
export enum MoviesList {
  Popular = 'popular',
  NowPlaying = 'now_playing',
  TopRated = 'top_rated',
  TrendingDay = 'trending_day',
  TrendingWeek = 'trending_week',
}

export enum TrendingPeriod {
  Day = 'TrendingDay',
  Week = 'TrendingWeek'}


export enum PosterSize {
  W92 = "w92",
  W154 = "w154",
  W185 = "w185",
  W342 = "w342",
  W500 = "w500",
  W780 = "w780",
  Original = "original",
}

export enum BackdropSize {
  W300 = "w300",
  W780 = "w780",
  W1280 = "w1280",
  Original = "original",
}

export enum Category {
  TRENDING = 'trending',
  POPULAR = 'popular',
  TOP_RATED = 'toprated',
  NOW_PLAYING = 'nowplaying',
}

export {}